+++ src/lib/Started.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import Stopwatch from './Stopwatch.svelte';
-    import { now } from './stores';
+    import { int } from './stores';
 </script>
 
-<Stopwatch {...$$props} now="{$now}"/>
+<Stopwatch {...$$props} now="{$int}"/>
